Moved up BaseClasses to main branch
route, bus, and stop classes are put into a new package and will be moved to main
