+++ osp-bus-app/src/main/java/busAppCore/baseClass/Bus.java
@@ -1,4 +1,4 @@
-package busAppCore;
+package baseClass;
 
 import dataDisplay.DisplayableObject;
 import dataDisplay.ListItemData;
