+++ osp-bus-app/src/main/java/busAppCore/baseClass/Route.java
@@ -1,4 +1,4 @@
-package busAppCore;
+package baseClass;
 
 import dataDisplay.DisplayableObject;
 import dataDisplay.ListItemData;
