+++ osp-bus-app/src/main/java/busAppCore/baseClass/Stop.java
@@ -1,4 +1,4 @@
-package busAppCore;
+package baseClass;
 
 import dataDisplay.*;
